+++ ZSZAdminWeb/Models/RoleEditModel.cs
@@ -10,6 +10,6 @@ namespace ZSZAdminWeb.Models
         public int Id { get; set; }
         public string Name { get; set; }
 
-        public int[] PermissionIds { get; set; }
+        public long[] PermissionIds { get; set; }
     }
 }
